align ReceiveStrategy names to transaction modes
